esercizio completato

+++ RandomTools/waittime.c
@@ -1,5 +1,5 @@
 #include <time.h>
-void WaitSeconds(int secondi)
+void WaitSeconds(float secondi)
 {
     time_t now = clock();
     for(; (clock() - now) < (secondi * CLOCKS_PER_SEC) ;) // Waits for 3 seconds
